+++ test/annotations/optional_test.dart
@@ -3,7 +3,7 @@ library example.optional;
 import 'package:ng_di/ng_di.dart';
 import 'package:test/test.dart';
 
-part 'optional_test.ng_di.g.dart';
+part 'optional_test.g.dart';
 
 @injectable
 class Engine {